Created a presentational component to make richlinks easier to snapshot
